+++ NEWS.md
@@ -8,6 +8,8 @@ Version 1.1.2 - beta
 - Release notes: https://ezdxf.mozman.at/release-v1-1.html
 - CHANGE: [#936](https://github.com/mozman/ezdxf/issues/936)
   improve modelspace extents updates
+- BUGFIX: [#939](https://github.com/mozman/ezdxf/issues/939)
+  `Matplotlib` requires oriented outer paths and holes to draw correct filled paths 
 
 Version 1.1.1 - 2023-10-08
 --------------------------